@@ -11,7 +11,7 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.12.0-dev.68.0 <3.0.0"
 
 dependencies:
   flutter:
