@@ -5,7 +5,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 void main() => runApp(const MyApp());
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key key}) : super(key: key);
+  const MyApp({Key? key}) : super(key: key);
 
   @override
   MyAppState createState() => MyAppState();
@@ -93,75 +93,19 @@ class MyAppState extends State<MyApp> {
         GlobalWidgetsLocalizations.delegate,
       ],
       home: Scaffold(
+        backgroundColor: Colors.red,
         appBar: AppBar(
           title: const Text('CountryPicker Example'),
         ),
         body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              CountryCodePicker(
-                onChanged: print,
-                // Initial selection and favorite can be one of code ('IT') OR dial_code('+39')
-                initialSelection: 'IT',
-                favorite: const ['+39', 'FR'],
-                countryFilter: const ['IT', 'FR'],
-                showFlagDialog: false,
-                comparator: (a, b) => b.name.compareTo(a.name),
-                //Get the country information relevant to the initial selection
-                onInit: (code) => debugPrint(
-                    "on init ${code.name} ${code.dialCode} ${code.name}"),
-              ),
-              CountryCodePicker(
-                onChanged: print,
-                // Initial selection and favorite can be one of code ('IT') OR dial_code('+39')
-                initialSelection: 'IT',
-                favorite: const ['+39', 'FR'],
-                countryFilter: const ['IT', 'FR'],
-                // flag can be styled with BoxDecoration's `borderRadius` and `shape` fields
-                flagDecoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(7),
-                ),
-              ),
-              const SizedBox(
-                width: 400,
-                height: 60,
-                child: CountryCodePicker(
-                  onChanged: print,
-                  hideMainText: true,
-                  showFlagMain: true,
-                  showFlag: false,
-                  initialSelection: 'TF',
-                  hideSearch: true,
-                  showCountryOnly: true,
-                  showOnlyCountryWhenClosed: true,
-                  alignLeft: true,
-                ),
-              ),
-              SizedBox(
-                width: 400,
-                height: 60,
-                child: CountryCodePicker(
-                  onChanged: (element) => debugPrint(element.toLongString()),
-                  initialSelection: 'TF',
-                  showCountryOnly: true,
-                  showOnlyCountryWhenClosed: true,
-                  favorite: const ['+39', 'FR'],
-                ),
-              ),
-              SizedBox(
-                width: 100,
-                height: 60,
-                child: CountryCodePicker(
-                  enabled: false,
-                  onChanged: (c) => c.name,
-                  initialSelection: 'TF',
-                  showCountryOnly: true,
-                  showOnlyCountryWhenClosed: true,
-                  favorite: const ['+39', 'FR'],
-                ),
-              ),
-            ],
+          child: CountryCodePicker(
+            barrierColor: Colors.transparent,
+            backgroundColor: Colors.transparent,
+            dialogBackgroundColor: Colors.transparent,
+            onChanged: print,
+            showFlagDialog: false,
+            onInit: (code) => debugPrint(
+                "on init ${code!.name} ${code.dialCode} ${code.name}"),
           ),
         ),
       ),
